+++ apps/web/CHANGELOG.md
@@ -1,5 +1,16 @@
 # tenno-companion
 
+## 0.2.1
+
+### Patch Changes
+
+- Fix minerva-velimir file name
+  Fix chatroom selector scrollbar
+  Fix chatroom selector active link
+  Remove unused string from translation files
+- Updated dependencies
+  - @tenno-companion/kim@0.1.1
+
 ## 0.2.0
 
 ### Minor Changes